// CRUD-App/src/models/transaction.ts
export interface Transaction {
    id: string;
    amount: number;
    description: String;
    date: string;
    year: number;
    month: number;
    // day: number;
    // time: string;
}

// const sample:Transaction[] = [
// {id: 'someid1', 
// amount: 100, 
// date: 'Thu Dec 28 2023 20:19:20',
// description: 'some iphone'},
// {id: 'someid2', 
// amount: 420, 
// date: 'Thu Dec 28 2023 20:19:22',
// description: 'some ipad'},
// {id: 'someid3', 
// amount: 629.00, 
// date: 'Thu Dec 28 2023 20:19:23',
// description: 'some macbook'},
// ];